Properly encode precerts in /get-entries
Turns out I completely misread a portion of the RFC, and was sending back
complete jibberish in the extra_data response key.  Fixing that up required
changing the way that /add-chain wrote entries to the queue, as well as what
process_queue did to those entries afterwards, so there was enough
information available to be able to give the right into to /get-entries.